move EntityFinder & EntityCollection to Entity\Query namespace
