+++ lib/Storage/EntityDatabase.php
@@ -8,8 +8,8 @@
 
 namespace DevNet\Entity\Storage;
 
-use DevNet\Entity\Internal\EntityFinder;
 use DevNet\Entity\Metadata\EntityModel;
+use DevNet\Entity\Query\EntityFinder;
 use DevNet\Entity\Query\EntityQueryProvider;
 use DevNet\Entity\Storage\EntityDataPersister;
 use DevNet\Entity\Tracking\EntityEntry;
